Reorganize project structure around dist build
