+++ .gitignore
@@ -8,9 +8,11 @@ pnpm-debug.log*
 lerna-debug.log*
 
 node_modules
-dist
 dist-ssr
 *.local
+
+# Build artifacts
+*.js.map
 
 # Editor directories and files
 !.vscode/extensions.json
