--- docs/index.css
@@ -1,1 +1,0 @@
-:root{font-family:-apple-system,BlinkMacSystemFont,Segoe UI,Roboto,Oxygen,Ubuntu,Cantarell,sans-serif;line-height:1.5;font-weight:400;--bg-primary: #ffffff;--bg-secondary: #f8f9fa;--text-primary: #212529;--text-secondary: #6c757d;--border-color: #dee2e6;--input-bg: #ffffff;--input-border: #ced4da;--input-focus: #0d6efd;--button-primary: #0d6efd;--button-primary-hover: #0b5ed7;--button-secondary: #6c757d;--button-secondary-hover: #5c636a;--error-color: #dc3545;--success-color: #198754;--card-shadow: 0 .125rem .25rem rgba(0, 0, 0, .075);font-synthesis:none;text-rendering:optimizeLegibility;-webkit-font-smoothing:antialiased;-moz-osx-font-smoothing:grayscale}[data-theme=dark]{--bg-primary: #1a1a1a;--bg-secondary: #2d2d2d;--text-primary: #e9ecef;--text-secondary: #adb5bd;--border-color: #495057;--input-bg: #2d2d2d;--input-border: #495057;--input-focus: #4d9fff;--button-primary: #4d9fff;--button-primary-hover: #3d8fef;--button-secondary: #6c757d;--button-secondary-hover: #5c636a;--card-shadow: 0 .125rem .25rem rgba(0, 0, 0, .3)}*{box-sizing:border-box;margin:0;padding:0}body{margin:0;min-height:100vh;background-color:var(--bg-secondary);color:var(--text-primary);display:flex;justify-content:center;align-items:center;padding:1rem}#app{width:100%;max-width:600px}.payment-form-container{background-color:var(--bg-primary);border-radius:.5rem;box-shadow:var(--card-shadow);overflow:hidden}.form-header{padding:1.5rem;border-bottom:1px solid var(--border-color);display:flex;justify-content:space-between;align-items:center}.form-header h1{font-size:1.5rem;font-weight:600;margin:0}.theme-toggle{background:none;border:1px solid var(--border-color);border-radius:.375rem;padding:.5rem;cursor:pointer;color:var(--text-primary);font-size:1.25rem;display:flex;align-items:center;justify-content:center;transition:all .2s;width:2.5rem;height:2.5rem}.theme-toggle:hover{background-color:var(--bg-secondary)}.theme-toggle:focus{outline:2px solid var(--input-focus);outline-offset:2px}.progress-bar{display:flex;padding:1rem 1.5rem;gap:.5rem;background-color:var(--bg-secondary)}.progress-step{flex:1;display:flex;flex-direction:column;align-items:center;gap:.5rem}.progress-circle{width:2rem;height:2rem;border-radius:50%;background-color:var(--input-bg);border:2px solid var(--border-color);display:flex;align-items:center;justify-content:center;font-weight:600;font-size:.875rem;color:var(--text-secondary);transition:all .3s}.progress-step.active .progress-circle{background-color:var(--button-primary);border-color:var(--button-primary);color:#fff}.progress-step.completed .progress-circle{background-color:var(--success-color);border-color:var(--success-color);color:#fff}.progress-label{font-size:.75rem;color:var(--text-secondary);text-align:center}.progress-step.active .progress-label{color:var(--text-primary);font-weight:500}.form-body{padding:2rem 1.5rem}.form-step{display:none;animation:fadeIn .3s ease-in}.form-step.active{display:block}@keyframes fadeIn{0%{opacity:0;transform:translateY(10px)}to{opacity:1;transform:translateY(0)}}.form-group{margin-bottom:1.25rem}.form-row{display:grid;grid-template-columns:1fr 1fr;gap:1rem}.form-row-3{display:grid;grid-template-columns:1fr 1fr minmax(9ch,12ch);gap:1rem}@media (max-width: 640px){.form-row,.form-row-3{grid-template-columns:1fr}}label{display:block;margin-bottom:.375rem;font-weight:500;font-size:.875rem;color:var(--text-primary)}label .required{color:var(--error-color);margin-left:.125rem}input,select{width:100%;padding:.625rem .75rem;font-size:1rem;line-height:1.5;color:var(--text-primary);background-color:var(--input-bg);border:1px solid var(--input-border);border-radius:.375rem;transition:border-color .15s ease-in-out,box-shadow .15s ease-in-out;font-family:inherit}input:focus,select:focus{outline:0;border-color:var(--input-focus);box-shadow:0 0 0 .2rem #0d6efd40}[data-theme=dark] input:focus,[data-theme=dark] select:focus{box-shadow:0 0 0 .2rem #4d9fff40}input::placeholder{color:var(--text-secondary);opacity:.7}input.error,select.error{border-color:var(--error-color)}input.error:focus,select.error:focus{box-shadow:0 0 0 .2rem #dc354540}.error-message{color:var(--error-color);font-size:.75rem;margin-top:.25rem;display:none}.error-message.show{display:block}.hint{color:var(--text-secondary);font-size:.75rem;margin-top:.25rem}.checkbox-group{display:flex;align-items:flex-start;gap:.5rem;margin-top:1rem}.checkbox-group input[type=checkbox]{width:auto;margin-top:.25rem;cursor:pointer}.checkbox-group label{margin-bottom:0;cursor:pointer;-webkit-user-select:none;user-select:none}#billing-address-fields{margin-top:1rem;padding-top:1rem;border-top:1px solid var(--border-color);display:none}#billing-address-fields.show{display:block}.card-input-group{display:grid;grid-template-columns:1fr 1fr;gap:1rem}.form-actions{display:flex;gap:1rem;margin-top:2rem;padding-top:1.5rem;border-top:1px solid var(--border-color)}button{padding:.625rem 1.25rem;font-size:1rem;font-weight:500;font-family:inherit;border:none;border-radius:.375rem;cursor:pointer;transition:all .2s;display:flex;align-items:center;justify-content:center;gap:.5rem}button:disabled{opacity:.5;cursor:not-allowed}button:focus{outline:2px solid var(--input-focus);outline-offset:2px}.btn-primary{background-color:var(--button-primary);color:#fff;flex:1}.btn-primary:hover:not(:disabled){background-color:var(--button-primary-hover)}.btn-secondary{background-color:transparent;color:var(--text-secondary);border:1px solid var(--border-color)}.btn-secondary:hover:not(:disabled){background-color:var(--bg-secondary);color:var(--text-primary)}.confirmation-content{text-align:center}.confirmation-icon{width:4rem;height:4rem;background-color:var(--success-color);border-radius:50%;display:flex;align-items:center;justify-content:center;margin:0 auto 1.5rem;color:#fff;font-size:2rem}.confirmation-content h2{font-size:1.5rem;margin-bottom:.5rem;color:var(--text-primary)}.confirmation-content p{color:var(--text-secondary);margin-bottom:2rem}.summary-section{background-color:var(--bg-secondary);border-radius:.375rem;padding:1rem;margin-bottom:1.5rem;text-align:left}.summary-section h3{font-size:.875rem;font-weight:600;text-transform:uppercase;color:var(--text-secondary);margin-bottom:.75rem}.summary-item{display:flex;justify-content:space-between;padding:.5rem 0;border-bottom:1px solid var(--border-color)}.summary-item:last-child{border-bottom:none}.summary-label{color:var(--text-secondary);font-size:.875rem}.summary-value{color:var(--text-primary);font-weight:500;font-size:.875rem}.loading-spinner{border:3px solid var(--border-color);border-top:3px solid var(--button-primary);border-radius:50%;width:1.25rem;height:1.25rem;animation:spin .8s linear infinite;display:none}.loading .loading-spinner{display:block}@keyframes spin{0%{transform:rotate(0)}to{transform:rotate(360deg)}}input[type=number]::-webkit-inner-spin-button,input[type=number]::-webkit-outer-spin-button{-webkit-appearance:none;margin:0}input[type=number]{-moz-appearance:textfield}
