--- docs/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Domain Payment Form</title>
-    <script type="module" crossorigin src="/index.js"></script>
-    <link rel="stylesheet" crossorigin href="/index.css">
-  </head>
-  <body>
-    <div id="app"></div>
-  </body>
-</html>
