+++ index.html
@@ -2,12 +2,13 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" href="data:," />
+    <link rel="stylesheet" href="./dist/main.css" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Domain Payment Form</title>
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/main.js"></script>
+    <script type="module" src="./dist/main.js"></script>
   </body>
 </html>
