+++ package.json
@@ -2,10 +2,9 @@
   "name": "payment-form",
   "version": "0.0.1",
   "scripts": {
-    "dev": "vite",
-    "build": "vite build --config vite.prod.config.js",
-    "build:dev": "vite build --outDir docs",
-    "preview": "vite preview --outDir docs",
+    "dev": "esbuild src/main.js --bundle --outdir=dist --format=esm --sourcemap --loader:.html=text --watch --servedir=.",
+    "build": "esbuild src/main.js --bundle --outdir=dist --format=esm --minify --loader:.html=text",
+    "preview": "npx serve dist",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:run": "vitest run"
@@ -28,7 +27,7 @@
   "license": "MIT",
   "devDependencies": {
     "jsdom": "^25.0.1",
-    "vite": "^7.1.9",
+    "esbuild": "^0.24.0",
     "vitest": "^3.2.4"
   }
 }
